% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSRPrisma_convert.R
\name{LSRPrisma_convert}
\alias{LSRPrisma_convert}
\title{Convert Data for LSR Prisma Flow Diagram between formats}
\usage{
LSRPrisma_convert(data, format = "approach1")
}
\arguments{
\item{data}{Unprocessed data provided in accordance with the
'template.csv' file.}

\item{format}{The desired output approach based on the four
potential layout formats for living systematic review PRISMA
flow diagrams. Options are: 'approach1' with each update
provided as a separate column; 'approach2' with the base and
all updates combined into one column; 'approach3' with the
base and all updates as two separate columns; and 'approach4'
with the base plus previous updates and the final update
provided as two separate columns.}
}
\value{
List of objects needed by \code{LSRPRISMA_flow()}.
}
\description{
Function to convert input data uploaded from format
used in Approach 1 (all updates added as separate columns) to
one of Approaches 2 to 4. Columns are summarised in a single
column (Approach 2), base plus updates (Approach 3), or all
updates and latest update (Approach 4). The function wraps
\code{LSRPrisma_data()} so that data can be directly passed to the
flow diagram creation function.
}
\examples{
\dontrun{
data <- read.csv('inst/extdata/approach1.csv')
flowdata <- LSRPrisma_convert(data, format = 'approach4')

}
}
