% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSRPrisma_flow.R
\name{LSRPrisma_flow}
\alias{LSRPrisma_flow}
\title{Living systematic review flow diagrams}
\usage{
LSRPrisma_flow(data, previous = TRUE, other = TRUE)
}
\arguments{
\item{data}{A dataset ready for plotting, produced as a list
of items using \code{LSRPrisma_data()}.}

\item{previous}{A logical argument specifying whether the 'previous
studies/report' arm should be included in the flow diagram (TRUE) or
not (FALSE).}

\item{other}{A logical argument specifying whether the 'other
sources' arm should be included in the flow diagram (TRUE) or
not (FALSE).}
}
\value{
A LSR Prisma plot of the data plotted onto a flow
diagram.
}
\description{
Produce various formats of PRISMA 2020 compliant living
systematic review flow diagrams with options for showing inputs from
previous reviews and other sources.
}
\examples{
\dontrun{
data <- read.csv('inst/extdata/approach1.csv')
data <- LSRPrisma_data(data)
plot <- LSRPrisma_flow(data, previous = FALSE, other = TRUE)
plot
}
}
