% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSRPrisma_convert.R
\name{process_excludes}
\alias{process_excludes}
\title{Process exclusion data}
\usage{
process_excludes(data)
}
\arguments{
\item{data}{Exclusions reasons imported in the template format.}
}
\value{
A single string of exclusions reasons and n formatted
for plotting.
}
\description{
Function converts raw exclusion reasons from separate
columns into a single string, involving consolidation across
duplicated reasons.
}
